requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ArXiv API
arxiv==2.0.0

# HTTP client
requests==2.31.0
urllib3==2.1.0

# AI/ML APIs
huggingface-hub==0.19.4
transformers==4.36.0

# Utilities
python-dateutil==2.8.2
click==8.1.7

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
